backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
pydantic-settings==2.10.1
passlib==1.7.4
PyYAML==6.0.2
itsdangerous==2.2.0
python-multipart==0.0.20
pytest==8.4.1
httpx==0.28.1
